+++ src/main/first-run.ts
@@ -1,7 +1,8 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { app, dialog } from 'electron';
-import log from 'electron-log';
+
+import { logError } from '../shared/logger';
 
 export async function onFirstRunMaybe() {
   if (isFirstRun()) {
@@ -49,7 +50,7 @@ function isFirstRun() {
 
     fs.writeFileSync(configPath, '');
   } catch (err) {
-    log.error('First run: Unable to write firstRun file', err);
+    logError('isFirstRun', 'Unable to write firstRun file', err);
   }
 
   return true;
