+++ src/main/updater.ts
@@ -2,6 +2,8 @@ import log from 'electron-log';
 import { autoUpdater } from 'electron-updater';
 import type { Menubar } from 'menubar';
 import { updateElectronApp } from 'update-electron-app';
+
+import { logError, logInfo } from '../shared/logger';
 import type MenuBuilder from './menu';
 
 export default class Updater {
@@ -18,29 +20,29 @@ export default class Updater {
     });
 
     autoUpdater.on('checking-for-update', () => {
-      log.info('Auto Updater: Checking for update');
+      logInfo('auto updater', 'Checking for update');
       this.menuBuilder.setCheckForUpdatesMenuEnabled(false);
     });
 
-    autoUpdater.on('error', (error) => {
-      log.error('Auto Updater: error checking for update', error);
+    autoUpdater.on('error', (err) => {
+      logError('auto updater', 'Error checking for update', err);
       this.menuBuilder.setCheckForUpdatesMenuEnabled(true);
     });
 
     autoUpdater.on('update-available', () => {
-      log.info('Auto Updater: New update available');
+      logInfo('auto updater', 'New update available');
       menuBuilder.setUpdateAvailableMenuEnabled(true);
       this.menubar.tray.setToolTip('Gitify\nA new update is available');
     });
 
     autoUpdater.on('update-downloaded', () => {
-      log.info('Auto Updater: Update downloaded');
+      logInfo('auto updater', 'Update downloaded');
       menuBuilder.setUpdateReadyForInstallMenuEnabled(true);
       this.menubar.tray.setToolTip('Gitify\nA new update is ready to install');
     });
 
     autoUpdater.on('update-not-available', () => {
-      log.info('Auto Updater: update not available');
+      logInfo('auto updater', 'Update not available');
       this.menuBuilder.setCheckForUpdatesMenuEnabled(true);
     });
   }
