+++ src/renderer/hooks/useNotifications.test.ts
@@ -2,6 +2,7 @@ import { act, renderHook, waitFor } from '@testing-library/react';
 import axios, { AxiosError } from 'axios';
 import nock from 'nock';
 
+import * as logger from '../../shared/logger';
 import {
   mockAuth,
   mockGitHubCloudAccount,
@@ -13,7 +14,6 @@ import {
   mockSingleNotification,
 } from '../utils/api/__mocks__/response-mocks';
 import { Errors } from '../utils/errors';
-import * as logger from '../utils/logger';
 import { useNotifications } from './useNotifications';
 
 describe('renderer/hooks/useNotifications.ts', () => {
