+++ src/renderer/routes/LoginWithOAuthApp.tsx
@@ -2,6 +2,8 @@ import { BookIcon, PersonIcon, SignInIcon } from '@primer/octicons-react';
 import { type FC, useCallback, useContext } from 'react';
 import { Form, type FormRenderProps } from 'react-final-form';
 import { useNavigate } from 'react-router-dom';
+
+import { logError } from '../../shared/logger';
 import { Header } from '../components/Header';
 import { Button } from '../components/buttons/Button';
 import { FieldInput } from '../components/fields/FieldInput';
@@ -21,7 +23,6 @@ import {
   isValidToken,
 } from '../utils/auth/utils';
 import { Constants } from '../utils/constants';
-import { logError } from '../utils/logger';
 
 interface IValues {
   hostname?: Hostname;
