+++ src/renderer/utils/helpers.ts
@@ -4,13 +4,14 @@ import {
   ChevronRightIcon,
 } from '@primer/octicons-react';
 import { formatDistanceToNowStrict, parseISO } from 'date-fns';
+
+import { logError, logWarn } from '../../shared/logger';
 import { defaultSettings } from '../context/App';
 import type { Chevron, Hostname, Link, SettingsState } from '../types';
 import type { Notification } from '../typesGitHub';
 import { getHtmlUrl, getLatestDiscussion } from './api/client';
 import type { PlatformType } from './auth/types';
 import { Constants } from './constants';
-import { logError, logWarn } from './logger';
 import {
   getCheckSuiteAttributes,
   getLatestDiscussionComment,
